+++ web/modules/custom/filmfolk_fixtures/src/Fixture/UserFixture.php
@@ -17,6 +17,10 @@ final class UserFixture extends AbstractFixture implements DependentFixtureInter
    */
   #[\Override]
   public function load() {
+
+    // Get the pre-created media entity using the reference key we set.
+    $default_profile_picture_media = $this->getReference('media:profile_picture:default');
+
     $user = User::create([
       'mail' => 'user@example.com',
       'field_navn' => 'Navn Navnesen',
@@ -28,6 +32,10 @@ final class UserFixture extends AbstractFixture implements DependentFixtureInter
       'field_description' => [
         'value' => '<strong>Hej</strong> med dig',
         'format' => 'simple',
+      ],
+      // Reference the media entity here.
+      'field_profile_picture' => [
+        'target_id' => $default_profile_picture_media->id(),
       ],
     ])
       ->activate();
@@ -181,6 +189,7 @@ final class UserFixture extends AbstractFixture implements DependentFixtureInter
   #[\Override]
   public function getDependencies() {
     return [
+      ProfilePictureFixture::class,
       KommuneTermFixture::class,
       FunktionTermFixture::class,
     ];